Renamed marked_cpairs to marked_pair_counting_engines
