+++ halotools/mock_observables/pair_counters/marked_pair_counting_engines/custom_weighting_func.pyx
