+++ halotools/mock_observables/pair_counters/marked_pair_counting_engines/marked_cylindrical_isolation_engine.pyx
