+++ halotools/mock_observables/pair_counters/marked_pair_counting_engines/setup_package.py
